#6625 – Fix circular dependency issue for Jest
