+++ packages/ketcher-core/src/application/editor/shared/utils.types.ts
@@ -1,3 +1,5 @@
-import { Bond } from 'domain/entities';
+import { Bond } from 'domain/entities/bond';
 
 export type BondAtoms = Pick<Bond, 'begin' | 'end'>;
+
+export type FlipDirection = 'horizontal' | 'vertical';
